backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
passlib[bcrypt]
bcrypt<5
pydantic[email]
python-jose[cryptography]